Split sandbox e2e tests
