+++ .github/workflows/health_checks.yml
@@ -262,6 +262,49 @@ jobs:
           aws-region: us-west-2
       - name: Run e2e deployment tests
         run: npm run test:dir packages/integration-tests/lib/test-e2e/deployment.test.js
+  e2e_sandbox:
+    if: needs.do_include_e2e.outputs.run_e2e == 'true'
+    strategy:
+      # will finish running other test matrices even if one fails
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-14-xlarge, windows-latest]
+        node-version: [18, 20]
+        # skip multiple node version test on other os
+        exclude:
+          - os: macos-14-xlarge
+            node-version: 20
+          - os: windows-latest
+            node-version: 20
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: ${{ matrix.os == 'windows-latest' && 35 || 25 }}
+    needs:
+      - do_include_e2e
+      - build
+    permissions:
+      # these permissions are required for the configure-aws-credentials action to get a JWT from GitHub
+      id-token: write
+      contents: read
+    steps:
+      - uses: actions/checkout@0ad4b8fadaa221de15dcec353f45205ec38ea70b # version 4.1.4
+      - uses: ./.github/actions/setup_node
+        with:
+          node-version: ${{ matrix.node-version }}
+      - uses: ./.github/actions/restore_build_cache
+      - run: cd packages/cli && npm link
+      - name: Configure test tooling credentials
+        uses: ./.github/actions/setup_profile
+        with:
+          role-to-assume: ${{ secrets.E2E_TOOLING_ROLE_ARN }}
+          aws-region: us-west-2
+          profile-name: e2e-tooling
+      - name: Configure test execution credentials
+        uses: aws-actions/configure-aws-credentials@e3dd6a429d7300a6a4c196c26e071d42e0343502 # version 4.0.2
+        with:
+          role-to-assume: ${{ secrets.E2E_RUNNER_ROLE_ARN }}
+          aws-region: us-west-2
+      - name: Run e2e sandbox tests
+        run: npm run test:dir packages/integration-tests/lib/test-e2e/sandbox.test.js
   e2e_backend_output:
     if: needs.do_include_e2e.outputs.run_e2e == 'true'
     runs-on: ubuntu-latest
@@ -487,6 +530,7 @@ jobs:
       - test_with_coverage
       - e2e_package_manager
       - e2e_deployment
+      - e2e_sandbox
       - e2e_create_amplify
     runs-on: ubuntu-latest
     steps:
