+++ packages/integration-tests/src/test-e2e/deployment.test.ts
@@ -8,16 +8,12 @@ import fs from 'fs/promises';
 import { shortUuid } from '../short_uuid.js';
 import { getTestProjectCreators } from '../test-project-setup/test_project_creator.js';
 import { TestProjectBase } from '../test-project-setup/test_project_base.js';
-import { userInfo } from 'os';
 import { PredicatedActionBuilder } from '../process-controller/predicated_action_queue_builder.js';
 import { ampxCli } from '../process-controller/process_controller.js';
 import path from 'path';
 import {
-  ensureDeploymentTimeLessThan,
   interruptSandbox,
   rejectCleanupSandbox,
-  replaceFiles,
-  waitForConfigUpdateAfterDeployment,
 } from '../process-controller/predicated_action_macros.js';
 import assert from 'node:assert';
 import { TestBranch, amplifyAppPool } from '../amplify_app_pool.js';
@@ -26,10 +22,6 @@ import { ClientConfigFormat } from '@aws-amplify/client-config';
 import { testConcurrencyLevel } from './test_concurrency.js';
 import { TestCdkProjectBase } from '../test-project-setup/cdk/test_cdk_project_base.js';
 import { getTestCdkProjectCreators } from '../test-project-setup/cdk/test_cdk_project_creator.js';
-import {
-  amplifySharedSecretNameKey,
-  createAmplifySharedSecretName,
-} from '../shared_secret.js';
 
 const testProjectCreators = getTestProjectCreators();
 const testCdkProjectCreators = getTestCdkProjectCreators();
@@ -107,81 +99,6 @@ void describe('deployment tests', { concurrency: testConcurrencyLevel }, () => {
               format
             );
           }
-        });
-      });
-    });
-
-    testProjectCreators.forEach((testProjectCreator) => {
-      void describe(`sandbox deploys ${testProjectCreator.name}`, () => {
-        let testProject: TestProjectBase;
-        let sandboxBackendIdentifier: BackendIdentifier;
-
-        before(async () => {
-          testProject = await testProjectCreator.createProject(rootTestDir);
-          sandboxBackendIdentifier = {
-            type: 'sandbox',
-            namespace: testProject.name,
-            name: userInfo().username,
-          };
-        });
-
-        after(async () => {
-          await testProject.tearDown(sandboxBackendIdentifier);
-        });
-
-        void describe('in sequence', { concurrency: false }, () => {
-          const sharedSecretsEnv = {
-            [amplifySharedSecretNameKey]: createAmplifySharedSecretName(),
-          };
-          void it(`[${testProjectCreator.name}] deploys fully`, async () => {
-            await testProject.deploy(
-              sandboxBackendIdentifier,
-              sharedSecretsEnv
-            );
-            await testProject.assertPostDeployment(sandboxBackendIdentifier);
-          });
-
-          void it('generates config after sandbox --once deployment', async () => {
-            const processController = ampxCli(
-              ['sandbox', '--once'],
-              testProject.projectDirPath,
-              {
-                env: sharedSecretsEnv,
-              }
-            );
-            await processController
-              .do(waitForConfigUpdateAfterDeployment())
-              .run();
-
-            await testProject.assertPostDeployment(sandboxBackendIdentifier);
-          });
-
-          void it(`[${testProjectCreator.name}] hot-swaps a change`, async () => {
-            const updates = await testProject.getUpdates();
-            if (updates.length > 0) {
-              const processController = ampxCli(
-                ['sandbox', '--dirToWatch', 'amplify'],
-                testProject.projectDirPath,
-                {
-                  env: sharedSecretsEnv,
-                }
-              );
-
-              for (const update of updates) {
-                processController
-                  .do(replaceFiles(update.replacements))
-                  .do(ensureDeploymentTimeLessThan(update.deployThresholdSec));
-              }
-
-              // Execute the process.
-              await processController
-                .do(interruptSandbox())
-                .do(rejectCleanupSandbox())
-                .run();
-
-              await testProject.assertPostDeployment(sandboxBackendIdentifier);
-            }
-          });
         });
       });
     });
